refactor: rewrite in ECMAScript notation cjsToEsm
