+++ src/modules/cjsToEsm.mjs
@@ -1,0 +1,36 @@
+import path from 'path';
+import * as fs from "fs/promises";
+import { fileURLToPath } from 'url';
+import { release, version } from 'os';
+import { createServer as createServerHttp } from 'http';
+import './files/c.js';
+
+const __filename = fileURLToPath(import.meta.url);
+const __dirname = path.dirname(__filename);
+const filesPath = path.join(__dirname, 'files');
+const a = JSON.parse(await fs.readFile(path.join(filesPath, 'a.json'), 'utf8'));
+const b = JSON.parse(await fs.readFile(path.join(filesPath, 'b.json'), 'utf8'));
+const random = Math.random();
+let unknownObject;
+
+if (random > 0.5) {
+    unknownObject = a;
+} else {
+    unknownObject = b;
+}
+
+console.log(`Release ${release()}`);
+console.log(`Version ${version()}`);
+console.log(`Path segment separator is "${path.sep}"`);
+
+console.log(`Path to current file is ${__filename}`);
+console.log(`Path to current directory is ${__dirname}`);
+
+const createMyServer = createServerHttp((_, res) => {
+    res.end('Request accepted');
+});
+
+export {
+    unknownObject,
+    createMyServer,
+};
